+++ src/ewp_waveform/__init__.py
@@ -1,3 +1,3 @@
-"""ewp-waveform package scaffold."""
+"""ewp-waveform application package."""
 
 __version__ = "0.0.0"